pyproject.toml
[tool.poetry]
name = "zdx905"
version = "0.1.0"
description = ""
authors = ["Liuwei <zhiwei_liu@dsa-auto.com.tw>"]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.8"
python-multipart = "^0.0.6"
uvicorn = "^0.23.2"
fastapi = "^0.104.1"
websockets = "^12.0"
PyJWT = "^2.8.0"
pyodbc = "^5.0.1"
SQLAlchemy = "^2.0.22"
alembic = "^1.12.1"
bcrypt = "^4.0.1"
shortuuid = "^1.0.11"

[tool.poetry.dev-dependencies]
requests = "^2.31.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
